Add BaseController
