+++ InvoiceMaker/InvoiceMaker/InvoiceMaker.csproj
@@ -102,6 +102,7 @@
     <Reference Include="System.EnterpriseServices" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Controllers\BaseController.cs" />
     <Compile Include="Controllers\WorksDoneController.cs" />
     <Compile Include="Controllers\WorkTypesController.cs" />
     <Compile Include="Data\Context.cs" />
